+++ R/PlotAnnotation.R
@@ -1,90 +1,136 @@
-add_text_plot <- function(
-    output_data,
-    text,
-    cal_age,
-    height,
-    plot_pretty = TRUE) {
+#' Add Text Annotation to Various Summary Plots
+#'
+#' @description
+#' Having plotted the output of the various modelling approaches, add text annotation at a given location on the plot. It is based on the base [text] function
+#'
+#' This function/annotation can be applied after having called any of [carbondate::PlotPosteriorMean]
+#'
+#' @inheritParams graphics::text
+#'
+#' @param output_plot The plot onto which you wish to add text
+#'
+#' @return None
+#'
+#' @export
+#'
+#' @examples
+#' # NOTE: This example is shown with a small n_iter and n_posterior_samples
+#' # to speed up execution.
+#' # Try n_iter and n_posterior_samples as the function defaults.
+#'
+#' pp_output <- PPcalibrate(
+#'     pp_uniform_phase$c14_age,
+#'     pp_uniform_phase$c14_sig,
+#'     intcal20,
+#'     n_iter = 1000,
+#'     show_progress = FALSE)
+#'
+#' # Default plot with 2 sigma interval
+#' posterior_mean_plot <- PlotPosteriorMeanRate(
+#'     pp_output,
+#'     n_posterior_samples = 100)
+#'
+#' # Add text to plot
+#' AddTextPlot(posterior_mean_plot,
+#'     x = 600, y = 500,
+#'     labels = expression(paste("600", ""^14, "C ", "yrs BP")),
+#'     cex = 0.7,
+#'     col = "red")
+#'
+AddTextPlot <- function(
+    output_plot,
+    x,
+    y,
+    labels,
+    adj = NULL,
+    pos = NULL,
+    offset = 0.5,
+    vfont = NULL,
+    cex = 1,
+    col = NULL,
+    font = NULL) {
 
   # Ensure revert to main environment par on exit of function
   oldpar <- graphics::par(no.readonly = TRUE)
   on.exit(graphics::par(oldpar))
 
-  # Set nice plotting parameters
-  if(plot_pretty) {
-    graphics::par(
-      new = TRUE,
-      mgp = c(3, 0.7, 0),
-      xaxs = "i",
-      yaxs = "i",
-      mar = c(5, 4.5, 4, 2) + 0.1,
-      las = 1)
-  }
-
-  xlim <- rev(range(two_normals_DPMM[[1]]$calendar_age_BP))
-  ylim <- c(0, 3 * max(two_normals_DPMM[[1]]$density_mean))
-
+  graphics::par(output_plot$plot_par)
+  text(x, y, labels, adj, pos, offset, vfont, cex, col, font)
 }
 
 
-# Store the old plotting parameters so you can change them back at the end
-oldpar <- par(no.readonly = TRUE)
+#' Add Shading to Various Summary Plots
+#'
+#' @description
+#' Having plotted the output of the various modelling approaches, add a shaded interval or box to the plot. This function/annotation can be applied after having called any of [carbondate::PlotPosteriorMean]
+#'
+#' @param output_plot The plot onto which you wish to add text
+#' @param x_start,x_end The values on the x-axis (calendar age) at which you want the shading to start and end
+#' @param y_start,y_end `NULL` if you want the shading to cover the entire y-axis range. Otherwise (if you want to plot a box) the y-axis values where you want the sahding to start and end
+#' @param col The color to be used
+#' @param alpha The level of transparency for the shading (`0 < alpha < 1`) to see the plot underneath. `0` means fully transparent and `1` means fully opaque.
+#'
+#' @return None
+#'
+#' @export
+#'
+#' @examples
+#' # NOTE: This example is shown with a small n_iter and n_posterior_samples
+#' # to speed up execution.
+#' # Try n_iter and n_posterior_samples as the function defaults.
+#'
+#' pp_output <- PPcalibrate(
+#'     pp_uniform_phase$c14_age,
+#'     pp_uniform_phase$c14_sig,
+#'     intcal20,
+#'     n_iter = 1000,
+#'     show_progress = FALSE)
+#'
+#' # Default plot with 2 sigma interval
+#' posterior_mean_plot <- PlotPosteriorMeanRate(
+#'     pp_output,
+#'     n_posterior_samples = 100)
+#'
+#' # Add transparent red shaded period to plot
+#' AddShadingPlot(posterior_mean_plot,
+#'     x_start = 620, x_end = 600,
+#'     col = "red")
+#'
+#' # Add green (more opaque) shaded box to plot
+#' AddShadingPlot(posterior_mean_plot,
+#'     x_start = 590, x_end = 550,
+#'     y_start = 500, y_end = 410,
+#'     col = "green", alpha = 0.8)
+#'
+AddShadingPlot <- function(
+    output_plot,
+    x_start,
+    x_end,
+    y_start = NULL,
+    y_end = NULL,
+    col, alpha = 0.4) {
 
-# Run a Polya Urn DPMM model (also works for Poisson process)
+  # Ensure revert to main environment par on exit of function
+  oldpar <- graphics::par(no.readonly = TRUE)
+  on.exit(graphics::par(oldpar))
 
-polya_urn_output <- PolyaUrnBivarDirichlet(
-  rc_determinations = two_normals$c14_age,
-  rc_sigmas = two_normals$c14_sig,
-  calibration_curve=intcal20,
-  n_iter = 1e4,
-  show_progress = FALSE)
+  add.alpha <- function(cols, alpha) {
+    grDevices::rgb(t(grDevices::col2rgb(cols)/255),
+                   alpha = alpha)
+  }
 
-two_normals_DPMM <- PlotPredictiveCalendarAgeDensity(
-  output_data = polya_urn_output,
-  show_SPD = TRUE)
+  shading_col <- add.alpha(col, alpha = alpha)
 
-# Adjust the plotting parameters to those used internally in the DPMM plotting function
-# Note: the ylim corresponds to the density (not the 14C age)
-par(new = TRUE,
-    mgp = c(3, 0.7, 0),
-    xaxs = "i",
-    yaxs = "i",
-    mar = c(5, 4.5, 4, 2) + 0.1,
-    las = 1)
-xlim <- rev(range(two_normals_DPMM[[1]]$calendar_age_BP))
-ylim <- c(0, 3 * max(two_normals_DPMM[[1]]$density_mean))
+  graphics::par(output_plot$plot_par)
+  if(is.null(y_start)) y_start <- output_plot$plot_par$usr[3]
+  if(is.null(y_end)) y_end <- output_plot$plot_par$usr[4]
 
-# Create an empty plot
-plot(
-  NULL,
-  NULL,
-  type = "n",
-  ylim = ylim,
-  xlim = xlim,
-  axes = FALSE,
-  xlab = NA,
-  ylab = NA,
-  xaxs = "i",
-  yaxs = "i")
-
-# Add calendar age ticks wherever you want
-xticks_minor <- seq(2000, 6000, by = 100)
-axis(1, xticks_minor, labels = FALSE, tcl = -0.2)
-
-# Add shading wherever you want
-period_start <- 4500
-period_end <- 4300
-
-# Decide on shading colour (and make somewhat transparent)
-shade_cols <- "red"
-add.alpha <- function(cols, alpha) rgb(t(col2rgb(cols)/255), alpha = alpha)
-cols <- add.alpha(shade_cols, alpha = 0.4)
-
-polygon(x = c(rep(period_start, 2), rep(period_end, 2)),
-        y = c(0, 0.82, 0.82, 0), border = NA,
-        col = cols)
-
-# Place 1/5 up image by using y = 0.2 * max(ylim)
-text("Shaded Period", x = period_end, y = 0.2 * max(ylim), pos = 4)
-
-# Reset/Change back to old plotting parameters
-par(oldpar)
+  graphics::polygon(x = c(rep(x_start, 2),
+                          rep(x_end, 2)),
+                    y = c(y_start,
+                          y_end,
+                          y_end,
+                          y_start),
+                    border = NA,
+                    col = shading_col)
+}